+++ plugins/triple_oscillator/TripleOscillator.h
@@ -66,8 +66,8 @@ private:
 	float m_volumeRight;
 
 	// normalized detuning -> x/sampleRate
-	float m_detuningLeft;
-	float m_detuningRight;
+	double m_detuningLeft;
+	double m_detuningRight;
 	// normalized offset -> x/360
 	float m_phaseOffsetLeft;
 	float m_phaseOffsetRight;
